requirements.txt
fastapi
kafka-python
langchain-core
langchain-google-genai
langgraph
pydantic
redis
uvicorn
